String normalization to unsigned chars, fixing warnings.

+++ src/intf/utils.h
@@ -5,7 +5,7 @@
 void itoah(uint64_t value, char* buffer);
 void memset(uint8_t* addr, uint8_t value, uint64_t count);
 void memcpy(uint8_t* src, uint8_t* dest, uint64_t count);
-void str_cpy(uint8_t* src, uint8_t* dest, uint8_t count);
+void str_cpy(uint8_t* src, uint8_t* dest, uint32_t count);
 
 int printf(const char* format, ...);
 int printf_ll(const char* format, ...);
